prints all the elements of a list_t list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	size_t count = 0;
+	int count = 0;
 
 	while (h)
 	{
@@ -18,7 +18,7 @@ size_t print_list(const list_t *h)
 			printf("[0] (nil)\n");
 		} else
 		{
-			printf("[%u] %s\n", h->len, h->str);
+			printf("[%d] %s\n", h->len, h->str);
 		}
 		count++;
 		h = h->next;
